requirements.txt
aerich==0.9.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
asyncclick==8.3.0.4
asyncpg==0.30.0
bcrypt==4.0.1
black==25.9.0
certifi==2025.10.5
click==8.3.0
Deprecated==1.2.14
dictdiffer==0.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
Faker==37.8.0
fastapi==0.118.0
googleapis-common-protos==1.59.1
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
iso8601==2.1.0
isort==6.1.0
loguru==0.7.3
mypy==1.18.2
mypy_extensions==1.1.0
opentelemetry-api==1.37.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
protobuf==4.25.8
psutil==6.1.0
pyasn1==0.6.1
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pypika-tortoise==0.6.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
pytokens==0.1.10
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
thrift==0.22.0
tortoise-orm==0.25.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.37.0
wrapt==1.17.3
zipp==3.23.0